code cleanup #98
