+++ src/Edit.c
@@ -2726,128 +2726,6 @@ void EditMoveDown(HWND hwnd)
   }
 }
 
-extern BOOL bAutoIndent;
-
-LPSTR GetLinePrefix(int iLine, LPBOOL pbLineEmpty)
-{
-  *pbLineEmpty = TRUE;
-  LPSTR pszPrefix = NULL;
-
-  const int iCurPos = SendMessage(hwndEdit, SCI_GETCURRENTPOS, 0, 0);
-  const int iLineStart = SendMessage(hwndEdit, SCI_POSITIONFROMLINE, iLine, 0);
-  const int iLinePrefixLength = iCurPos - iLineStart;
-  pszPrefix = GlobalAlloc(GPTR, iLinePrefixLength + 1);
-  struct TextRange tr = { { iLineStart, iCurPos }, pszPrefix };
-  SendMessage(hwndEdit, SCI_GETTEXTRANGE, 0, (LPARAM)&tr);
-
-  for (size_t i = 0; i < strlen(pszPrefix); ++i)
-  {
-    const char chCurrent = pszPrefix[i];
-    if (!isspace(chCurrent) || (chCurrent == '\r') || (chCurrent == '\n'))
-    {
-      pszPrefix[i] = 0;
-      *pbLineEmpty = (i == 0);
-      break;
-    }
-  }
-
-  return pszPrefix;
-}
-
-void FreeLinePrefix(LPSTR pszPrefix)
-{
-  if (pszPrefix)
-  {
-    GlobalFree(pszPrefix);
-  }
-}
-
-void InsertNewLineWithPrefix(LPSTR pszPrefix, BOOL bInsertAbove)
-{
-  const BOOL bAutoIndentOrigin = bAutoIndent;
-  bAutoIndent = 0;
-  SendMessage(hwndEdit, SCI_NEWLINE, 0, 0);
-  bAutoIndent = bAutoIndentOrigin;
-  if (bInsertAbove)
-  {
-    SendMessage(hwndEdit, SCI_CHARLEFT, 0, 0);
-  }
-  if (pszPrefix && (strlen(pszPrefix) > 0))
-  {
-    const int iCurrentPos = SendMessage(hwndEdit, SCI_GETCURRENTPOS, 0, 0);
-    SendMessage(hwndEdit, SCI_INSERTTEXT, iCurrentPos, (LPARAM)pszPrefix);
-    SendMessage(hwndEdit, SCI_LINEEND, 0, 0);
-  }
-}
-
-void EditInsertNewLine(HWND hwnd, BOOL insertAbove)
-{
-  if (HLS_Edit_selection_stop(HL_SE_APPLY))
-  {
-    return;
-  }
-
-  const int iCurPos = SendMessage(hwnd, SCI_GETCURRENTPOS, 0, 0);
-  const int iCurLine = SendMessage(hwnd, SCI_LINEFROMPOSITION, iCurPos, 0);
-  const int iCurLineEndPos = SendMessage(hwnd, SCI_GETLINEENDPOSITION, iCurLine, 0);
-  SendMessage(hwndEdit, SCI_BEGINUNDOACTION, 0, 0);
-  const int iPrevLine = (iCurLine > 0) ? iCurLine - 1 : 0;
-  LPSTR pszPrefixText = NULL;
-  BOOL bIsEmptyPrefix = FALSE;
-
-  if (insertAbove)
-  {
-    const int iPrevLineEndPos = (iPrevLine == iCurLine) ? 0 : SendMessage(hwnd, SCI_GETLINEENDPOSITION, iPrevLine, 0);
-    if (bAutoIndent)
-    {
-      const int iLineStart = SendMessage(hwnd, SCI_POSITIONFROMLINE, iCurLine, 0);
-      const int iLinePrefixLength = iCurPos - iLineStart;
-      bIsEmptyPrefix = (iLinePrefixLength == 0);
-      if (!bIsEmptyPrefix)
-      {
-        pszPrefixText = GetLinePrefix(iCurLine, &bIsEmptyPrefix);
-      }
-      if (bIsEmptyPrefix)
-      {
-        SendMessage(hwndEdit, SCI_SETSEL, iPrevLineEndPos, iPrevLineEndPos);
-        SendMessage(hwndEdit, SCI_NEWLINE, 0, 0);
-        if (iPrevLine == iCurLine)
-        {
-          const int iNewPrevLineEndPos = SendMessage(hwnd, SCI_GETLINEENDPOSITION, iPrevLine, 0);
-          SendMessage(hwndEdit, SCI_SETSEL, iNewPrevLineEndPos, iNewPrevLineEndPos);
-        }
-      }
-      else
-      {
-        SendMessage(hwndEdit, SCI_SETSEL, iPrevLineEndPos, iPrevLineEndPos);
-        InsertNewLineWithPrefix(pszPrefixText, (iPrevLine == iCurLine));
-      }
-    }
-    else
-    {
-      SendMessage(hwndEdit, SCI_SETSEL, iPrevLineEndPos, iPrevLineEndPos);
-      SendMessage(hwndEdit, SCI_NEWLINE, 0, 0);
-    }
-  }
-  else
-  {
-    const BOOL isLineEnd = (iCurPos == iCurLineEndPos);
-    if (isLineEnd && bAutoIndent)
-    {
-      SendMessage(hwndEdit, SCI_SETSEL, iCurLineEndPos, iCurLineEndPos);
-      pszPrefixText = GetLinePrefix(iCurLine-1, &bIsEmptyPrefix);
-      InsertNewLineWithPrefix(pszPrefixText, FALSE);
-    }
-    else
-    {
-      SendMessage(hwndEdit, SCI_SETSEL, iCurLineEndPos, iCurLineEndPos);
-      SendMessage(hwndEdit, SCI_NEWLINE, 0, 0);
-    }
-  }
-  SendMessage(hwndEdit, SCI_ENDUNDOACTION, 0, 0);
-  FreeLinePrefix(pszPrefixText);
-}
-
 //=============================================================================
 //
 //  EditModifyLines()
